update as per comments
