+++ querybook/webapp/lib/utils/cron.ts
@@ -1,3 +1,4 @@
+import * as Yup from 'yup';
 import moment from 'moment';
 
 export interface IRecurrence {
@@ -151,3 +152,29 @@ export function validateCronForRecurrrence(cron: string) {
 
     return true;
 }
+
+export const recurrenceOnYup = Yup.object().when(
+    'recurrence',
+    (recurrence: string, schema) => {
+        const onSchema: any = {};
+        if (recurrence === 'weekly') {
+            onSchema.dayWeek = Yup.array()
+                .min(1)
+                .of(Yup.number().min(0).max(6))
+                .required();
+        } else {
+            onSchema.dayMonth = Yup.array()
+                .min(1)
+                .of(Yup.number().min(1).max(31))
+                .required();
+        }
+
+        if (recurrence === 'yearly') {
+            onSchema.month = Yup.array()
+                .min(1)
+                .of(Yup.number().min(1).max(12))
+                .required();
+        }
+        return Yup.object().shape(onSchema);
+    }
+);
